Former-commit-id: 965bad2a1cc9e3c08a81399f7a34e613bd89e258

+++ unisim_lib/unisim/service/debug/gdb_server/gdb_server.hh
@@ -175,7 +175,7 @@ public:
 
 private:
 	static const unsigned int MAX_BUFFER_SIZE = 256;
-	bool ParseHex(const string& s, unsigned int& pos, ADDRESS& value);
+	bool ParseHex(const string& s, size_t& pos, ADDRESS& value);
 	bool GetChar(char& c, bool blocking);
 	bool PutChar(char c);
 	bool GetPacket(string& s, bool blocking);
